refactor: moves test dynamics module one level up
